fix, improving installer

+++ installer.sh
@@ -8,13 +8,26 @@ if [ `whoami` == "root" ]
 then
     $style;
     echo "Putting the rights to execute the script...";
-    chmod 755 battery;
-    echo "Moving script in /usr/bin";
-    mv -i makefile /usr/bin;
-    echo "Putting manpage in /usr/share/man/man1";
-    mv -i makefile.1.gz /usr/share/man/man1;
-    echo "Removing installer.sh";
-    rm -f installer.sh;
+    chmod 755 makefile;
+    echo "Moving script in /usr/bin...";
+    if [ "$1" == "-m" ]
+    then
+	mv -i makefile /usr/bin;
+    else
+	cp -i makefile /usr/bin;
+    fi
+    echo "Putting manpage in /usr/share/man/man1...";
+    if [ "$1" == "-m" ]
+    then
+	mv -i makefile.1.gz /usr/share/man/man1;
+    else
+	cp -i makefile.1.gz /usr/share/man/man1;
+    fi
+    if [ "$1" == "-m" ]
+    then
+	echo "Removing installer.sh..."
+	rm -f installer.sh;
+    fi
 else
     $red;
     echo "You need to be root to install correctly the battery script.";
